refactor(desktop): scope the sidebar to the route it serves
One rail served every route, so it had to carry the union of what every
route needed and hand the routes it did not fit the leftovers as dead
chrome. Sources, Outputs and Folders belong to a conversation but rendered
on home and settings, disabled, until any chat had been opened — after
which they stopped being disabled and started navigating back into
whichever conversation was open last.

That last part came from `ChatListStore.selected`, a mirror of the URL that
the chat route wrote on mount and nothing cleared on the way out. The shell
went on believing a chat was open while the reader stood on home.

The rail is now chosen by the route. Home gets the global controls, a
conversation gets its own and nothing else, and settings gets home's. The
mirrored selection is gone; the open chat is read from the URL, which
removes the class of bug rather than guarding against it. The searchable
chat list moves out of a panel beside one conversation and onto home,
where looking for a different conversation actually happens, and the
`chats` panel type goes with it — a stale URL naming it falls back to the
conversation alone.

Losing the chat list from the conversation's rail would also lose the
marker that another chat is waiting, so the way back reports it instead.

Tests: replaces the two Sidebar files with coverage of what the split is
for — that each route offers only controls it can act on, and that the
conversation's rail carries no chat list — and keeps the collapse-
persistence case the deleted files covered. The settings watcher test
drops its prompt-detail assertion, which encoded settings remembering a
chat id; the summary poll and the dock notification it actually cares
about are unchanged.

Closes #651
