+++ crates/openwave-desktop/ui/src/ChatExplorer.dom.test.tsx
@@ -1,11 +1,11 @@
 // @vitest-environment jsdom
 import { cleanup, screen } from "@testing-library/react";
-import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
+import { afterEach, beforeEach, describe, expect, it } from "vitest";
 
 import type { Chat } from "./api";
 import { useChatAttention } from "./ChatAttention";
 import { useChatListStore } from "./ChatListStore";
-import { ChatsPanel } from "./ChatsPanel";
+import { ChatExplorer } from "./ChatExplorer";
 import { renderWithRouter } from "./test/router";
 
 const chats: Chat[] = [
@@ -23,10 +23,10 @@ beforeEach(() => {
 });
 afterEach(cleanup);
 
-describe("ChatsPanel", () => {
+describe("ChatExplorer", () => {
   it("marks waiting chats throughout the searchable list", async () => {
     useChatAttention.getState().setChatIdsWithPendingPrompts(["chat-2"]);
-    await renderWithRouter(<ChatsPanel activeChatId={null} onNewChat={vi.fn()} />);
+    await renderWithRouter(<ChatExplorer />);
 
     expect(screen.getByLabelText("Release notes needs attention")).toBeInTheDocument();
     expect(screen.queryByLabelText("Roadmap needs attention")).not.toBeInTheDocument();
