+++ crates/openwave-desktop/ui/src/SettingsRoute.tsx
@@ -1,7 +1,8 @@
-import { useNavigate } from "@tanstack/react-router";
+import { useNavigate, useRouter } from "@tanstack/react-router";
 
 import { useApp } from "./AppContext";
-import { useChatListStore } from "./ChatListStore";
+import { RouteFrame } from "./RouteFrame";
+import { HomeSidebar } from "./sidebar/HomeSidebar";
 import { SettingsView } from "./SettingsView";
 
 /**
@@ -10,10 +11,16 @@ import { SettingsView } from "./SettingsView";
  * The conversation unmounts while this is open, which it can afford to: the
  * transcript is a view of a durable journal and rehydrates on the way back, and
  * the watcher that notices the agent asking a question lives in the shell, not
- * in the conversation. Going back returns to the chat that was open.
+ * in the conversation.
+ *
+ * Going back is the previous history entry rather than a remembered chat id.
+ * Settings is reachable from everywhere, so where the reader came from is the
+ * only answer that is right from all of them — and it is one fewer place that
+ * has to keep its own copy of which conversation is open.
  */
 export function SettingsRoute() {
   const navigate = useNavigate();
+  const router = useRouter();
   const {
     client,
     models,
@@ -25,24 +32,25 @@ export function SettingsRoute() {
     checkForUpdate,
     restartForUpdate,
   } = useApp();
-  const openChatId = useChatListStore((state) => state.selected?.id ?? null);
 
   return (
-    <SettingsView
-      client={client}
-      models={models}
-      providers={providers}
-      onProvidersChanged={() => void refreshCatalog()}
-      onBack={() => {
-        void (openChatId
-          ? navigate({ to: "/c/$chatId", params: { chatId: openChatId } })
-          : navigate({ to: "/" }));
-      }}
-      themeMode={themeMode}
-      onThemeChange={setThemeMode}
-      updateState={updateState}
-      onCheckForUpdate={checkForUpdate}
-      onRestartForUpdate={restartForUpdate}
-    />
+    <RouteFrame sidebar={<HomeSidebar />}>
+      <SettingsView
+        client={client}
+        models={models}
+        providers={providers}
+        onProvidersChanged={() => void refreshCatalog()}
+        onBack={() => {
+          // A deep link straight to settings has nothing behind it.
+          if (router.history.canGoBack()) router.history.back();
+          else void navigate({ to: "/" });
+        }}
+        themeMode={themeMode}
+        onThemeChange={setThemeMode}
+        updateState={updateState}
+        onCheckForUpdate={checkForUpdate}
+        onRestartForUpdate={restartForUpdate}
+      />
+    </RouteFrame>
   );
 }
